requirements.txt
astroid==2.6.6
autopep8==1.5.7
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.5
colorama==0.4.4
flake8==3.9.2
idna==3.2
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
libusb1==2.0.1
mccabe==0.6.1
ndeflib==0.3.3
nfcpy==1.0.3
parso==0.8.2
pluggy==1.0.0
pycodestyle==2.7.0
pyDes==2.0.1
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.9.6
pyserial==3.5
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.2
requests==2.26.0
rope==0.19.0
snowballstemmer==2.1.0
soupsieve==2.2.1
toml==0.10.2
ujson==4.1.0
urllib3==1.26.6
wrapt==1.12.1
yapf==0.31.0
